Select individual quests. Move notification base to proper folder
